requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
pdf2image==1.16.3
requests==2.31.0
boto3
